Criado as iterações da lista

+++ app/src/test/java/br/com/bg7/appvistoria/data/inspection/InspectionTestBase.java
@@ -197,7 +197,7 @@ class InspectionTestBase {
         }
 
         @Override
-        public void onProgressUpdated(Inspection inspection, double progress) {
+        public void onProgressUpdated(Inspection inspection, Integer progress) {
 
         }
